pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Lanngraph-Wave-Orchestrator"
version = "1.1.8"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.1.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.0.1",
    "langgraph-cli[inmem]>=0.1.0",

]

[tool.hatch.build.targets.wheel]
packages = ["langgraph_wave_orchestrator"]
